\name{sheffer.surv}
\alias{sheffer.surv}
\docType{data}
\title{
Survival information of Sheffer et al. colorectal dataset
}
\description{
Partial data from Sheffer et al. paper
}
\usage{data(sheffer.surv)}
\format{
  \describe{
    \item{\code{samples}}{the samples name in Sheffer data}
    \item{\code{time}}{Survival time for each sample}
    \item{\code{event}}{Survival status for each sample}
  }
}

\source{
Sheffer et.\ al. Association of survival and disease progression with chromosomal instability: A genomic exploration of colorectal cancer.
\emph{PNAS}, 2009, Vol 106(17) pp: 7131-7136.
}
\examples{
data(sheffer.surv)
}
\keyword{datasets}
